agents/pyproject.toml
[project]
name = "bedrock-claude4-agent"
version = "0.1.0"
description = "Bedrock Claude 4を使用したAIエージェント実装"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=0.1.0",
    "strands-agents>=1.0.1",
    "boto3>=1.0.0",
    "mcp>=1.12.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
